+++ database/migrations/2024_11_29_082403_create_worker_payments_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->date("transaction_date");
             $table->integer("week");
-            $table->integer('total');
+            $table->bigInteger('total');
             $table->string("attachment")->nullable();
             $table->unsignedBigInteger('project_id');
             $table->foreign('project_id')
